feat: 实现动态配置系统和基于角色的权限控制
- 添加数据库迁移：用户角色字段、应用设置表、邀请码表
- 实现配置系统核心功能，支持运行时动态配置管理
- 增强JWT声明包含用户角色信息
- 统一类型系统，所有ID使用i64类型
- 添加管理员中间件和配置管理API端点
- 实现完整的邀请码生成、验证和管理功能
- 增强用户注册验证逻辑，支持邮箱域名白名单、密码强度要求
- 更新种子数据，添加管理员和普通用户测试账号
- 添加配置审计日志表用于追踪配置变更

此功能将应用业务配置从静态文件迁移到数据库，实现基于角色的权限控制，
提供完整的动态配置管理和邀请码注册系统。

+++ src/api/services/oidc_service.rs
@@ -104,7 +104,7 @@ pub async fn userinfo(
         .ok_or(AppError::Unauthorized)?;
 
     // 解析 user_id
-    let user_id: i32 = claims
+    let user_id: i64 = claims
         .sub
         .parse()
         .map_err(|_| AppError::Internal("Invalid user_id in token".into()))?;
